+++ src/tools/tools.ts
@@ -21,6 +21,7 @@ import * as performanceTools from './performance.js';
 import * as screencastTools from './screencast.js';
 import * as screenshotTools from './screenshot.js';
 import * as scriptTools from './script.js';
+import * as serviceWorkerTools from './serviceWorker.js';
 import * as slimTools from './slim/tools.js';
 import * as snapshotTools from './snapshot.js';
 import * as storageTools from './storage.js';
@@ -46,6 +47,7 @@ export const createTools = (args: ParsedArguments) => {
         ...Object.values(screencastTools),
         ...Object.values(screenshotTools),
         ...Object.values(scriptTools),
+        ...Object.values(serviceWorkerTools),
         ...Object.values(snapshotTools),
         ...Object.values(storageTools),
         ...Object.values(webmcpTools),